requirements.txt
ecdsa==0.13
gevent==1.0.2
greenlet==0.4.9
paramiko==1.16.0
pbr==1.8.1
pyaml==15.8.2
pycrypto==2.6.1
PyGithub==1.26.0
PyYAML==3.11
voluptuous==0.8.7
wheel==0.24.0
